+++ onnxruntime/core/session/ort_env.cc
@@ -9,6 +9,7 @@
 #include "core/session/ort_apis.h"
 #include "core/session/environment.h"
 #include "core/session/allocator_adapters.h"
+#include "core/session/user_logging_sink.h"
 #include "core/common/logging/logging.h"
 #include "core/framework/provider_shutdown.h"
 #include "core/platform/logging/make_platform_default_log_sink.h"
@@ -19,17 +20,6 @@ using namespace onnxruntime::logging;
 std::unique_ptr<OrtEnv> OrtEnv::p_instance_;
 int OrtEnv::ref_count_ = 0;
 onnxruntime::OrtMutex OrtEnv::m_;
-
-LoggingWrapper::LoggingWrapper(OrtLoggingFunction logging_function, void* logger_param)
-    : logging_function_(logging_function), logger_param_(logger_param) {
-}
-
-void LoggingWrapper::SendImpl(const onnxruntime::logging::Timestamp& /*timestamp*/, const std::string& logger_id,
-                              const onnxruntime::logging::Capture& message) {
-  std::string s = message.Location().ToString();
-  logging_function_(logger_param_, static_cast<OrtLoggingLevel>(message.Severity()), message.Category(),
-                    logger_id.c_str(), s.c_str(), message.Message().c_str());
-}
 
 OrtEnv::OrtEnv(std::unique_ptr<onnxruntime::Environment> value1)
     : value_(std::move(value1)) {
@@ -50,8 +40,8 @@ OrtEnv* OrtEnv::GetInstance(const OrtEnv::LoggingManagerConstructionInfo& lm_inf
     std::unique_ptr<LoggingManager> lmgr;
     std::string name = lm_info.logid;
     if (lm_info.logging_function) {
-      std::unique_ptr<ISink> logger = std::make_unique<LoggingWrapper>(lm_info.logging_function,
-                                                                       lm_info.logger_param);
+      std::unique_ptr<ISink> logger = std::make_unique<UserLoggingSink>(lm_info.logging_function,
+                                                                        lm_info.logger_param);
       lmgr = std::make_unique<LoggingManager>(std::move(logger),
                                               static_cast<Severity>(lm_info.default_warning_level),
                                               false,
