+++ onnxruntime/core/session/ort_env.h
@@ -5,26 +5,14 @@
 #include <atomic>
 #include <string>
 #include "core/session/onnxruntime_c_api.h"
-#include "core/common/logging/isink.h"
 #include "core/platform/ort_mutex.h"
 #include "core/common/status.h"
+#include "core/common/logging/logging.h"
 #include "core/framework/allocator.h"
 
 namespace onnxruntime {
 class Environment;
 }
-
-class LoggingWrapper : public onnxruntime::logging::ISink {
- public:
-  LoggingWrapper(OrtLoggingFunction logging_function, void* logger_param);
-
-  void SendImpl(const onnxruntime::logging::Timestamp& /*timestamp*/, const std::string& logger_id,
-                const onnxruntime::logging::Capture& message) override;
-
- private:
-  OrtLoggingFunction logging_function_;
-  void* logger_param_;
-};
 
 struct OrtEnv {
  public:
